feat: modified global so that we can use if everywhere, split program in two files
